basic error handling on -r build option
