+++ lib/skpm-build.js
@@ -6,8 +6,7 @@ var program = require('commander')
 var chalk = require('chalk')
 var uniqBy = require('lodash.uniqby')
 var objectAssign = require('object-assign')
-var WebpackShellPlugin = require('webpack-shell-plugin')
-var config = require('./utils/config').get()
+var generateWebpackConfig = require('./utils/webpackConfig')
 
 var buildEmojis = ['🔧', '🔨', '⚒', '🛠', '⛏', '🔩']
 function randomBuildEmoji () {
@@ -51,31 +50,6 @@ try {
   process.exit(1)
 }
 
-var userDefinedWebpackConfig = {}
-try {
-  if (fs.existsSync(path.join(process.cwd(), 'webpack.config.js'))) {
-    userDefinedWebpackConfig = require(path.join(process.cwd(), 'webpack.config.js'))
-  }
-} catch (err) {
-  console.error(chalk.red('error') + ' Error while reading webpack.config.js')
-  console.error(err)
-  process.exit(1)
-}
-
-var babelrcPath = path.join(process.cwd(), '.babelrc')
-var userDefinedBabelConfig = null
-try {
-  if (fs.existsSync(babelrcPath)) {
-    userDefinedBabelConfig = JSON.parse(fs.readFileSync(babelrcPath, 'utf8'))
-  } else if (packageJSON.babel) {
-    userDefinedBabelConfig = packageJSON.babel
-  }
-} catch (err) {
-  console.error(chalk.red('error') + ' Error while reading babelrc')
-  console.error(err)
-  process.exit(1)
-}
-
 if (!fs.existsSync(output)) {
   fs.mkdirSync(output)
 }
@@ -90,86 +64,7 @@ if (!fs.existsSync(path.join(output, 'Contents', 'Sketch'))) {
 
 var manifestFolder = path.dirname(manifest)
 
-function babelLoader (userDefinedBabelConfig) {
-  var options = {
-    babelrc: false,
-    plugins: [require('babel-plugin-add-module-exports')]
-  }
-
-  if (userDefinedBabelConfig) {
-    options = objectAssign(options, userDefinedBabelConfig)
-  } else {
-    options = objectAssign(options, {
-      presets: [require('babel-preset-airbnb')]
-    })
-  }
-
-  return {
-    test: /\.jsx?$/,
-    exclude: /node_modules/,
-    use: {
-      loader: 'babel-loader',
-      options: options
-    }
-  }
-}
-
-function webpackConfig (file, commandIdentifier) {
-  var basename = path.basename(file)
-
-  var plugins = userDefinedWebpackConfig.plugins || []
-
-  if (commandIdentifier) {
-    plugins.push(
-      new webpack.ProvidePlugin({
-        'console': 'sketch-module-console-polyfill',
-        'fetch': 'sketch-module-fetch-polyfill'
-      })
-    )
-  }
-
-  if (commandIdentifier && program.run) {
-    plugins.push(
-      new WebpackShellPlugin({
-        onBuildEnd: [
-          config.sketchPath + '/Contents/Resources/sketchtool/bin/sketchtool run "' + output + '" ' + commandIdentifier
-        ],
-        dev: false
-      })
-    )
-  }
-
-  return objectAssign(
-    {
-      module: {
-        rules: [
-          babelLoader(userDefinedBabelConfig)
-        ]
-      },
-      resolve: {
-        extensions: ['.sketch.js', '.js']
-      },
-      resolveLoader: {
-        modules: [
-          'node_modules',
-          path.join(__dirname, '..', 'node_modules')
-        ]
-      }
-    },
-    userDefinedWebpackConfig,
-    {
-      entry: path.join(commandIdentifier ? manifestFolder : process.cwd(), file),
-      output: {
-        filename: basename,
-        library: commandIdentifier ? 'onRun' : undefined,
-        path: commandIdentifier
-          ? path.join(output, 'Contents', 'Sketch')
-          : path.join(output, 'Contents', 'Resources')
-      },
-      plugins: plugins
-    }
-  )
-}
+var webpackConfig = generateWebpackConfig(program, output, manifestFolder, packageJSON)
 
 function copyManifest () {
   return new Promise(function (resolve, reject) {
